requirements.txt
Flask==2.2.5
tensorflow==2.20.0
keras==3.10.0
Pillow==10.2.0
gunicorn==21.2.0
numpy==1.26.4
